+++ CHANGELOG.md
@@ -15,9 +15,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 - Added option to use EdgeSplit modifier when importing model
 - Auto exit edit mode when exporting
 - Set correct smoothing based on material when importing
+- Automatically add .tga to material names on export if not present
+- Replace unsupported characters on export with '_'
+- Empty 'floor_level' object in the scene now defines in-game floor level
 ### Changed
 ### Fixed
 - Fixed unnecessary info printing on export
+- If object have no materials colwhite.tga will be used by default
 
 ## [1.4.1] - 2020-07-04
 ### Added